Fixed instance var type inference for initialized types (#96)
This fixes a small bug where type was not being inferred from code such as:

    var storedVariableWithConstructedImplicitType = Bool(booleanLiteral: true)

Problem was that the `inferType(..)` code was trying to split on the first `(` character, but was also including `()` in groups to ignore. I think it's safe to do the split only excluding <> and [] groups, can't think of any other case where a `(` would not delineate the type from its init params (tuples are handled by earlier code paths).