Output diagnostic warnings and errors that integrate with Xcode
- Generate "stub" mocks for protocols that have opaque inheritance, e.g.
  inherit from an externally-defined type without a supporting source
  file definition.
- Refactor the logging setup to better integrate with Xcode's by using
  `error:` and `warning:` prefixes that Xcode can parse.

We want the generator to be "nice" by default such that it shouldn't
(knowingly) create mocks that might fail compilation due to missing
supporting source files and the like.

Top-level types that can't be generated with full inheritance are now
minimally created and marked as unavailable.

In order to better communicate what the generator is doing, there's now
diagnostic warnings that integrate with Xcode’s buildtime issues. This
can be controlled by listing what types should be included (similar to
clang warnings) in the `--diagnostics` generator option.

By default, diagnostic warnings are _disabled_ as it could cause
confusion for new users and break test targets set to treat all warnings
as errors.

- `all` Emit all diagnostic warnings
- `not-mockable` Warn when skipping declarations that cannot be mocked
- `undefined-type` Warn on external types not defined in a supporting
  source file
- `type-inference` Warn when skipping complex property assignments in
  class mocks