Refactor type inference system for class properties
Previous type inference logic was largely copied from Sourcery, but has
a lot of issues with false positives and edge cases.

This is a very light refactor in that it specifically fixes regressions
and known bugs but does not attempt to do anything "smart" such as
parsing all declared symbols and using that for type inference.

For future releases we should pass in more info from the New Type System
into the type inferrer, although this may come at the cost of increased
parsing times.