Add support for reading JSON project descriptions (#171)
Custom build systems such as Buck or Bazel need a way to generate mocks
without an `.xcodeproj` file. This adds support for passing a JSON
project description containing all of the required module and dependency
graph information usually parsed from an Xcode project. The description
format is roughly compatible with the output from
`$ swift package describe --type JSON`, although adding the `dependencies`
field needs to be upstreamed.

```json
{
  "targets": [
    {
      "name": "MyLibrary",
      "type": "library",
      "path": "/path/to/my-library",
      "dependencies": [],
      "sources": [
        "SourceFileA.swift",
        "SourceFileB.swift"
      ]
    },
    {
      "name": "MyOtherLibrary",
      "type": "library",
      "path": "/path/to/my-other-library",
      "dependencies": [
        "MyLibrary"
      ],
      "sources": [
        "SourceFileA.swift",
        "SourceFileB.swift"
      ]
    },
    {
      "name": "MyLibraryTests",
      "type": "test",
      "path": "/path/to/my-library-tests",
      "dependencies": [
        "MyLibrary"
      ],
      "sources": [
        "SourceFileA.swift",
        "SourceFileB.swift"
      ]
    }
  ]
}
```

The `--project` generator option now accepts arbitrary files. Iff a file
with an `xcodeproj` extension is provided then it is read as an Xcode
project. Cache metadata is stored in the same enclosing directory as the
project description file under the `.mockingbird` subdirectory.

Since builds are run outside of an Xcode testing context, it's necessary
to optionally provide a host test target with the `--testbundle` option
to enable thunk pruning support.

```bash
mockingbird generate \
  --target MyLibrary \
  --testbundle MyLibraryTests \
  --project /path/to/project-desc.json
```