[4/n] Add thunk pruning method generator option (#174)
For certain projects it's unacceptable for source changes to break
builds due to malformed generated testing code. Because even generating
declarations for thunk stubs results in a large surface area for
unexpected compiler errors, it's necessary to provide a way to opt into
safer code generation.

This change adds a pruning method option to the generator with the
ability to completely omit code for unused mocked types. Note that
although it is possible for mocked types referenced in tests to regress
from source changes, the likelihood is much smaller.

The downside of omitting unused types is that writing tests will require
building the test bundle whenever a new mocked type is referenced in
order to run the generator and bootstrap the declarations. However, the
`mock` function will show an inline error in Xcode if the type hasn't
been generated, so the devX isn't too bad here.

In some respects, this is an admission that getting to that "perfect"
subset of the Swift compiler which can handle 99% of the code out there
just isn't feasible. We're maybe at 90% today, but the effort required
in getting the last 9% probably better spent in improving other areas of
the framework.