Add versioned CLI launcher
Most devs currently don’t pin the CLI because it’s not part of the
“happy” path in the setup instructions. Devs that do decide to pin the
CLI usually do so by checking it into source control which is not
sustainable long term.

This change makes pinning the binary part of the happy path of setting
up the framework for all package managers by adding a “launcher” to
abstract the CLI initialization process. The launcher allows the CLI to
be downloaded on-demand from an artifact provider. By default, GitHub’s
release artifacts is used, but it’s possible to set the download URL at
various levels:

- `REPO_URL` for artifacts hosted on a forked GitHub repo
- `ARTIFACTS_URL` for versioned artifacts at /<version>/Mockingbird.zip
- `ZIP_RELEASE_URL` for artifacts at arbitrary URLs

A side effect of the launcher is that it’s now trivial to run against an
arbitrary CLI version simply by setting the `MKB_VERSION` env var.