Fix generator caching for multi-project setups
Test target caching is currently done relative to the source project
which causes improper invalidation when the source and test targets
are in the same Xcode project. The downstream effect is that incremental
Xcode builds are much slower for large projects since the generated mock
file is “modified” each run.

This fixes the test target caching by using the environment-supplied
project and srcroot values for the cache key.