Fix configurator relative path handling (#264)
The configurator produces incorrect paths when run from an Xcode project
directory that differs from the test target project that’s being
configured. This PR fixes the issue by allowing the configurator to pass
a working directory to the generator so that argument values are
inferred from the correct base path.