Add Algolia DocSearch to the documentation (#271)
Searchability of the documentation is important but something that DocC
currently lacks. It’s possible to parse the generated JSON data files to
create a search index, but there aren’t any existing tools to do so. To
simplify integration for now, we can use Algolia DocSearch which is an
off-the-shelf, free, and hosted solution that’s used by popular projects
like React Native. In the future we should invest in decoupling the
search from Algolia, either by writing our own tool or using one that
integrates directly into DocC.

## Implementation

Although the integration is mostly straightforward, there are a few
caveats. The search button in the header is rendered inside a shadow
DOM, but DocSearch injects search modal into the document root. This
means that the stylesheet is “loaded” twice, once for the main document
and once for the shadow DOM. In practice, most browsers should merge and
cache the duplicate resource load, but this isn’t necessarily
guaranteed.

Another issue is that DocC uses its own data tag
`body[data-color-scheme]` to define dark and light themes, while
DocSearch expects it on `html[data-theme]`, which seems to be more
standard. A bit of JavaScript is required to synchronize theme changes
from DocC to DocSearch, including handling changes when using the “auto”
system-defined theme setting.