Enable mocking sources in test bundles (#286)
## Overview

It’s uncommon, but possible, to need to mock types defined in a test
bundle. For example, one use case is supplying a common set of default
implementations for partial mocking. A non-typical case is mocking types
in test bundle A and using them in test bundle B, which is
straightforward to set up in SwiftPM / JSON project descriptions but not
so much with an Xcode project.

To simplify the automatic setup process and reduce potential user
errors, the configurator will continue to validate that the
configuration target is a test bundle and each source target is not a
test bundle. Advanced users can modify the build phase manually to
bypass this as needed.

## Test Plan

Ran the generator against the `MockingbirdTests` target with various
thunk pruning levels.