Fix cross-project cache invalidation (#278)
## Overview

Test and source target caching is improperly invalidated when generating
mocks for targets outside of the current test target context. The
caching issues are compounded when generating mocks for multiple
projects in the same test target. This PR fixes the caching behavior for
cross-project mocking setups.

## Test Plan

Validated that the test and source target caches are invalidated iff the
inputs are changed.